Modification et restructuration des inputs, outputs et paramètre
