+++ src/classes/output.py
@@ -1,0 +1,6 @@
+class Output:
+    def __init__(self, led, sound, playerMove):
+        self.led=led
+        self.sound=sound
+        self.playerMove=playerMove
+    pass
