+++ src/classes/parameter.py
@@ -1,0 +1,8 @@
+class Parameter:
+    def __init__(self, light, mode, volume, level, reset):
+        self.light=light
+        self.mode=mode
+        self.volume=volume
+        self.level=level
+        self.reset=reset
+    pass
